requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Single-cell analysis
scanpy>=1.9.0
anndata>=0.9.0
scvi-tools>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML utilities
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
h5py>=3.8.0
loompy>=3.0.7

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
ipykernel>=6.23.0
jupyter>=1.0.0
